requirements.txt

torch
tiktoken
pyyaml
datasets
tqdm
